Split frame text cache helpers
